refactor(*): numeric types and better exceptions
