+++ package.json
@@ -68,18 +68,21 @@
   },
   "dependencies": {
     "@tensorflow/tfjs": "0.13.3",
+    "build": "^0.1.4",
     "isomorphic-fetch": "2.2.1",
     "libsvm-js": "0.2.0",
     "lodash": "4.17.10",
     "numeric": "1.2.6",
     "random-js": "1.0.8",
-    "stopword": "0.1.10"
+    "stopword": "0.1.10",
+    "yarn": "^1.15.2"
   },
   "devDependencies": {
     "@types/es6-promise": "3.3.0",
-    "@types/jest": "24.0.11",
+    "@types/jest": "^24.0.11",
     "@types/lodash": "4.14.108",
     "@types/node": "8.9.4",
+    "@types/numeric": "1.2.0",
     "@types/random-js": "1.0.30",
     "all-contributors-cli": "5.4.1",
     "del-cli": "1.1.0",
@@ -97,9 +100,9 @@
     "tslint": "5.4.3",
     "tslint-config-prettier": "1.8.0",
     "tslint-immutable": "4.5.1",
-    "typedoc": "0.11.1",
-    "typedoc-plugin-markdown": "1.1.13",
-    "typescript": "3.3.3333",
+    "typedoc": "0.14.2",
+    "typedoc-plugin-markdown": "1.1.27",
+    "typescript": "3.4.1",
     "vuepress": "0.12.0",
     "watch": "1.0.2"
   },
