+++ src/lib/naive_bayes/gaussian.ts
@@ -1,7 +1,8 @@
 import * as tf from '@tensorflow/tfjs';
 import { zip } from 'lodash';
-import { reshape, validateFitInputs, validateMatrix2D } from '../ops';
 import { IMlModel, Type1DMatrix, Type2DMatrix } from '../types';
+import { reshape } from '../utils/tensors';
+import { validateFitInputs, validateMatrix2D } from '../utils/validation';
 
 const SQRT_2PI = Math.sqrt(Math.PI * 2);
