--- src/lib/ops/index.ts
@@ -1,17 +1,0 @@
-import {
-  inferShape,
-  reshape,
-  validateFitInputs,
-  validateMatrix1D,
-  validateMatrix2D,
-  validateMatrixType
-} from './tensor_ops';
-
-export {
-  inferShape,
-  validateFitInputs,
-  validateMatrix1D,
-  validateMatrix2D,
-  validateMatrixType,
-  reshape
-};
