+++ src/lib/utils/Errors.ts
@@ -1,0 +1,3 @@
+export class ValidationError extends Error {}
+export class Validation1DMatrixError extends ValidationError {}
+export class Validation2DMatrixError extends ValidationError {}
