+++ test/Errors.ts
@@ -1,0 +1,7 @@
+export const validate1DMatrixErrorMessage =
+  'validateMatrix1D has received a non-array argument';
+export const validate2DMatrixErrorMessage =
+  'validateMatrix2D has received a non-array argument';
+export const tensorCreationErrorMessage =
+  'values passed to tensor(values) must be an array of numbers or booleans, or a TypedArray';
+export const matrixEmptyErrorMessage = 'Cannot fit with an empty value';
